Feat/wish

+++ src/main/java/com/tossclone/securities/dao/WishDao.java
@@ -9,10 +9,10 @@ import com.tossclone.securities.dto.Wish;
 @Mapper
 public interface WishDao {
 
-	public List<Wish> getWishList(int userId) throws Exception;
+	public List<Wish> getWishList(Long userId) throws Exception;
 
 	public void addWish(Wish wish) throws Exception;
 
-	public void deleteWish(int userId, String stockCode) throws Exception;
+	public void deleteWish(Long userId, String stockCode) throws Exception;
 
 }
